requirements.txt
# ── Core HTTP + parsing ──────────────────────────────────────────
requests>=2.31.0
urllib3>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0          # faster HTML parser for BeautifulSoup

# ── Config ───────────────────────────────────────────────────────
PyYAML>=6.0

# ── Google Sheets API ────────────────────────────────────────────
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.2.0

# ── Utility ──────────────────────────────────────────────────────
python-dateutil>=2.8.2
